Extracted live reload test into separate test module
Live reload tests must never execute in parallel with other tests since they will trigger a reload
in all open UIs, interfere with concurrently running tests. Having them in a separate test module
makes them more isolated. Also test with non-root context.
