+++ flow-tests/pom.xml
@@ -289,6 +289,7 @@
                 <module>test-subcontext</module>
                 <module>test-root-ui-context</module>
                 <module>test-router-custom-context</module>
+                <module>test-live-reload</module>
 
                 <module>test-scalability</module>
                 <module>test-memory-leaks</module>
