+++ flow-tests/test-live-reload/src/main/resources/META-INF/services/com.vaadin.flow.server.VaadinServiceInitListener
@@ -1,0 +1,1 @@
+com.vaadin.flow.uitest.ui.TestingServiceInitListener
