+++ flow-tests/test-live-reload/src/test/java/com/vaadin/flow/uitest/ui/LiveReloadIT.java
@@ -13,22 +13,19 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
+package com.vaadin.flow.uitest.ui;
 
-package com.vaadin.flow.uitest.ui.frontend;
-
-import net.jcip.annotations.NotThreadSafe;
 import java.util.List;
 
+import net.jcip.annotations.NotThreadSafe;
 import org.junit.Assert;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
-
-import com.vaadin.flow.testutil.ChromeBrowserTest;
+import org.openqa.selenium.html5.WebStorage;
 
 @NotThreadSafe
-public class LiveReloadIT extends ChromeBrowserTest {
+public class LiveReloadIT extends AbstractLiveReloadIT {
 
     @Test
     public void overlayShouldRender() {
@@ -59,24 +56,9 @@ public class LiveReloadIT extends ChromeBrowserTest {
     }
 
     @Test
-    @Ignore
     public void overlayShouldNotBeRenderedAfterDisable() {
         open();
-        waitForElementPresent(By.tagName("vaadin-devmode-gizmo"));
-        WebElement liveReload = findElement(By.tagName("vaadin-devmode-gizmo"));
-        liveReload.click();
-
-        WebElement liveReloadIcon = findInShadowRoot(liveReload,
-                By.className("gizmo")).get(0);
-        liveReloadIcon.click();
-
-        WebElement button = findInShadowRoot(liveReload, By.id("disable"))
-                .get(0);
-        button.click();
-
-        Assert.assertEquals(0,
-                findElements(By.tagName("vaadin-devmode-gizmo")).size());
-
+        disableLiveReloadInLocalStorage();
         driver.navigate().refresh();
 
         Assert.assertEquals(0,
@@ -84,22 +66,12 @@ public class LiveReloadIT extends ChromeBrowserTest {
     }
 
     @Test
-    @Ignore
     public void liveReloadShouldNotTriggerAfterDisable() {
         open();
-        waitForElementPresent(By.tagName("vaadin-devmode-gizmo"));
-        WebElement liveReload = findElement(By.tagName("vaadin-devmode-gizmo"));
-        liveReload.click();
+        disableLiveReloadInLocalStorage();
+        driver.navigate().refresh();
 
         String instanceId = findElement(By.id("elementId")).getText();
-
-        WebElement liveReloadIcon = findInShadowRoot(liveReload,
-                By.className("gizmo")).get(0);
-        liveReloadIcon.click();
-
-        WebElement button = findInShadowRoot(liveReload, By.id("disable"))
-                .get(0);
-        button.click();
 
         WebElement liveReloadTrigger = findElement(
                 By.id("live-reload-trigger-button"));
@@ -166,5 +138,10 @@ public class LiveReloadIT extends ChromeBrowserTest {
         Assert.assertFalse(
                 gizmo2.getAttribute("class").contains("active"));
         Assert.assertTrue(gizmo2.getAttribute("class").contains("gizmo"));
+    }
+
+    private void disableLiveReloadInLocalStorage() {
+        ((WebStorage) getDriver()).getLocalStorage()
+                .setItem("vaadin.live-reload.enabled", "false");
     }
 }
